Rename UD slice to E slice
The old UD coordinates were taking into account everything but the UD
slice. It was not wrong, but it could use a better naming. The reason
for this change is that when we want to get the actual sorted
coordinates for the 8 edges in the UD slice we get a namin conflict.
The E slice is the cube layer that separates the cube horizontaly,
spliting the U and D slice apart.